+++ lib/reviewer/setup/catalog.rb
@@ -122,7 +122,10 @@ module Reviewer
           commands: {
             review: 'bundle exec rake test'
           },
-          files: { review: 'bundle exec ruby -Itest', pattern: '*_test.rb', map_to_tests: 'minitest' },
+          files: {
+            review: "bundle exec ruby -Itest -e 'ARGV.each { |file| require File.expand_path(file) }'",
+            pattern: '*_test.rb', map_to_tests: 'minitest'
+          },
           detect: {
             gems: %w[minitest],
             directories: %w[test]